pyproject.toml
[project]
name = "a2a-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "crewai>=1.6.1",
    "langchain>=1.2.10",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
a2a-sdk = { git = "https://github.com/google/a2a-python.git" }
